Change snackbar duration to short. Make dark theme more contrast.

+++ app/src/main/java/com/github/kutyrev/poemmaster/ui/screens/mainlist/MainListScreen.kt
@@ -68,7 +68,7 @@ fun MainListScreen(
                                 it.poemHeader.name
                             ),
                             actionLabel = context.getString(R.string.snackbar_undo_button),
-                            duration = SnackbarDuration.Long
+                            duration = SnackbarDuration.Short
                         )
                         when (snackbarResult) {
                             SnackbarResult.Dismissed -> {
